# openapi/openapi.yaml
openapi: 3.0.3

info:
  title: Marketplace API
  version: 1.0.0
  description: API for the Marketplace application

servers:
  - url: http://localhost:5000

security: [ ]

paths:
  /products:
    get:
      operationId: listProducts
      summary: List products
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum number of products to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          required: false
          description: Opaque token used to retrieve the next page of products.
          schema:
            type: string
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'

        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      operationId: createProduct
      summary: Create a product
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: The same key with the same request body must return the same response.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProduct'
      responses:
        '201':
          description: Product created successfully
          headers:
            Idempotency-Replay:
              description: Indicates that the response was returned from a previous request with the same Idempotency-Key and request body.
              schema:
                type: boolean
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Idempotency key was reused with a different request body
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /orders:
    get:
      operationId: listOrders
      summary: List orders

      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum number of orders to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20

        - name: cursor
          in: query
          required: false
          description: Opaque token used to retrieve the next page of orders.
          schema:
            type: string

      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderList'

        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      operationId: createOrder
      summary: Create an order

      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: The same key with the same request body must return the same response.
          schema:
            type: string

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrder'

      responses:
        '201':
          description: Order created successfully
          headers:
            Idempotency-Replay:
              description: Indicates that the response was returned from a previous request with the same Idempotency-Key and request body.
              schema:
                type: boolean
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

        '404':
          description: Product or buyer not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

        '409':
          description: Product is out of stock or the buyer has insufficient funds
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

        '422':
          description: Idempotency key was reused with a different request body
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /orders/{id}:
    get:
      operationId: getOrder
      summary: Get an order by ID

      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the order.
          schema:
            type: string

      responses:
        '200':
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

        '404':
          description: Order not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  schemas:
    Product:
      type: object
      required:
        - id
        - name
        - price_cents
      properties:
        id:
          type: string
        name:
          type: string
        price_cents:
          type: integer
          minimum: 0
        stock:
          type: integer
          minimum: 0
    ProductList:
      type: object
      required:
        - items
        - next_cursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        next_cursor:
          type: string
          nullable: true
    CreateProduct:
      type: object
      additionalProperties: false
      required:
        - name
        - price_cents
      properties:
        name:
          type: string
        price_cents:
          type: integer
          minimum: 0
        description:
          type: string
        stock:
          type: integer
          minimum: 0
    OrderList:
      type: object
      required:
        - items
        - next_cursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        next_cursor:
          type: string
          nullable: true
    Order:
      type: object
      required:
        - id
        - created_at
        - status
        - items
        - total_cents
      properties:
        id:
          type: string
          description: Unique order identifier.
        created_at:
          type: string
          format: date-time
          description: Date and time when the order was created.
        status:
          type: string
          enum:
            - pending
            - paid
            - cancelled
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItem'
        total_cents:
          type: integer
          minimum: 0
          description: Total order amount in cents.
    OrderItem:
      type: object
      required:
        - product_id
        - quantity
        - price_cents
      properties:
        product_id:
          type: string
          description: Identifier of the ordered product.
        quantity:
          type: integer
          minimum: 1
          description: Number of products ordered.
        price_cents:
          type: integer
          minimum: 0
          description: Product price in cents at the time of ordering.
    CreateOrder:
      type: object
      required:
        - user_id
        - items
      properties:
        user_id:
          type: string
          description: Identifier of the buyer.
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateOrderItem'
    CreateOrderItem:
      type: object
      additionalProperties: false
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: string
        quantity:
          type: integer
          minimum: 1
    Problem:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - instance
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the type of problem.
        title:
          type: string
          description: Short human-readable summary of the problem.
        status:
          type: integer
          description: HTTP status code.
        detail:
          type: string
          description: Explanation specific to this occurrence of the problem.
        instance:
          type: string
          format: uri-reference
          description: URI identifying the specific occurrence of the problem.